RA-1174 - :hammer: Amending unit tests around wage unit.
