+++ src/Esfa.Vacancy.Register.UnitTests/Api/Mappings/ApprenticeshipMapper/WhenGettingLiveVacancyWithCustomRangeWageType.cs
@@ -11,16 +11,16 @@ namespace Esfa.Vacancy.Register.UnitTests.Api.Mappings.ApprenticeshipMapper
     [TestFixture]
     public class WhenGettingLiveVacancyWithCustomRangeWageType
     {
-        [TestCase(WageUnit.Weekly, 14000, 16000, "£14,000.00 - £16,000.00")]
-        [TestCase(WageUnit.Weekly, null, 16000, "Unknown - £16,000.00")]
-        [TestCase(WageUnit.Weekly, 14000, null, "£14,000.00 - Unknown")]
-        [TestCase(WageUnit.Monthly, 14000, 16000, "£14,000.00 - £16,000.00")]
-        [TestCase(WageUnit.Monthly, null, 16000, "Unknown - £16,000.00")]
-        [TestCase(WageUnit.Monthly, 14000, null, "£14,000.00 - Unknown")]
-        [TestCase(WageUnit.Annually, 14000, 16000, "£14,000.00 - £16,000.00")]
-        [TestCase(WageUnit.Annually, null, 16000, "Unknown - £16,000.00")]
-        [TestCase(WageUnit.Annually, 14000, null, "£14,000.00 - Unknown")]
-        public void ShouldHaveWageSetForVacanciesWithCustomRangeWageType(WageUnit wageUnit, decimal? lowerBound, decimal? upperBound, string expectedWageText)
+        [TestCase(2, WageUnit.Weekly, 14000, 16000, "£14,000.00 - £16,000.00")]
+        [TestCase(2, WageUnit.Weekly, null, 16000, "Unknown - £16,000.00")]
+        [TestCase(2, WageUnit.Weekly, 14000, null, "£14,000.00 - Unknown")]
+        [TestCase(3, WageUnit.Monthly, 14000, 16000, "£14,000.00 - £16,000.00")]
+        [TestCase(3, WageUnit.Monthly, null, 16000, "Unknown - £16,000.00")]
+        [TestCase(3, WageUnit.Monthly, 14000, null, "£14,000.00 - Unknown")]
+        [TestCase(4, WageUnit.Annually, 14000, 16000, "£14,000.00 - £16,000.00")]
+        [TestCase(4, WageUnit.Annually, null, 16000, "Unknown - £16,000.00")]
+        [TestCase(4, WageUnit.Annually, 14000, null, "£14,000.00 - Unknown")]
+        public void ShouldHaveWageSetForVacanciesWithCustomRangeWageType(int wageUnitId, WageUnit expectedWageUnit, decimal? lowerBound, decimal? upperBound, string expectedWageText)
         {
             const int weeklyWage = 2550;
             const int vacancyReference = 1234;
@@ -38,13 +38,13 @@ namespace Esfa.Vacancy.Register.UnitTests.Api.Mappings.ApprenticeshipMapper
                 .With(v => v.WageType, (int) WageType.CustomRange)
                 .With(v => v.WeeklyWage, weeklyWage)
                 .Without(v => v.WageText)
-                .With(v => v.WageUnitId, (int) wageUnit)
+                .With(v => v.WageUnitId, wageUnitId)
                 .Create();
 
             var vacancy = sut.MapToApprenticeshipVacancy(apprenticeshipVacancy);
 
             vacancy.VacancyReference.Should().Be(vacancyReference);
-            vacancy.WageUnit.Should().Be(wageUnit);
+            vacancy.WageUnit.Should().Be(expectedWageUnit);
             vacancy.WageText.Should().Be(expectedWageText);
         }
     }
